Upgrade sqlmesh to the latest
