+++ warehouse/metrics_mesh/models/events_daily_to_artifact.sql
