+++ warehouse/metrics_mesh/tests/test_events_daily_to_artifact.yml
@@ -11,43 +11,43 @@ test_events_daily_to_artifact:
         from_artifact_id: user_0
         event_source: BLOCKCHAIN
         event_type: CONTRACT_INVOCATION_SUCCESS_DAILY_COUNT
-        time: 2024-01-01T00:00:00Z
+        time: "2024-01-01 00:00:00+00:00"
         amount: 10
       - to_artifact_id: contract_1
         from_artifact_id: user_0
         event_source: BLOCKCHAIN
         event_type: CONTRACT_INVOCATION_SUCCESS_DAILY_COUNT
-        time: 2024-01-01T00:00:00Z
+        time: "2024-01-01 00:00:00+00:00"
         amount: 10
       - to_artifact_id: repo_0
         from_artifact_id: dev_0
         event_source: SOURCE_PROVIDER
         event_type: COMMIT_CODE
-        time: 2024-01-01T02:00:00Z
+        time: "2024-01-01 02:00:00+00:00"
         amount: 1
       - to_artifact_id: repo_0
         from_artifact_id: dev_0
         event_source: SOURCE_PROVIDER
         event_type: COMMIT_CODE
-        time: 2024-01-01T03:00:00Z
+        time: "2024-01-01 03:00:00+00:00"
         amount: 1
       - to_artifact_id: repo_0
         from_artifact_id: dev_0
         event_source: SOURCE_PROVIDER
         event_type: COMMIT_CODE
-        time: 2024-01-01T04:00:00Z
+        time: "2024-01-01 04:00:00+00:00"
         amount: 1
       - to_artifact_id: repo_0
         from_artifact_id: dev_1
         event_source: SOURCE_PROVIDER
         event_type: COMMIT_CODE
-        time: 2024-01-01T04:00:00Z
+        time: "2024-01-01 04:00:00+00:00"
         amount: 1
       - to_artifact_id: repo_0
         from_artifact_id: dev_1
         event_source: SOURCE_PROVIDER
         event_type: COMMIT_CODE
-        time: 2024-01-02T04:00:00Z
+        time: "2024-01-02 04:00:00+00:00"
         amount: 1
   outputs:
     query:
