+++ warehouse/metrics_tools/factory/__init__.py
@@ -1,4 +1,4 @@
 # ruff: noqa: F403
 
-from .factory import *
 from ..definition import *
+from .factory import *
